+++ xiaodianbao-shareintf/src/main/java/com/fengshang/xiaodianbao/uc/entity/WxUserExtra.java
@@ -1,7 +1,6 @@
 package com.fengshang.xiaodianbao.uc.entity;
 
 import com.xyz.tools.db.bean.BaseEntity;
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class WxUserExtra extends BaseEntity<Integer> {
@@ -17,9 +16,9 @@ public class WxUserExtra extends BaseEntity<Integer> {
 
     private String detailAddr;
 
-    private BigDecimal lng;
+    private Double lng;
 
-    private BigDecimal lat;
+    private Double lat;
 
     private Date createTime;
 
@@ -73,19 +72,19 @@ public class WxUserExtra extends BaseEntity<Integer> {
         this.detailAddr = detailAddr == null ? null : detailAddr.trim();
     }
 
-    public BigDecimal getLng() {
+    public Double getLng() {
         return lng;
     }
 
-    public void setLng(BigDecimal lng) {
+    public void setLng(Double lng) {
         this.lng = lng;
     }
 
-    public BigDecimal getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public void setLat(BigDecimal lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
